bit of refactoring for the project status stuff + minor wording tweaks
